+++ src/actions/index.js
@@ -1,0 +1,9 @@
+export const ADD_PRICE = "ADD_PRICE";
+export const addItem = item => {
+  return { type: ADD_PRICE, payload: item };
+};
+
+export const REMOVE_PRICE = "REMOVE_PRICE";
+export const removePrice = item => {
+  return { type: REMOVE_PRICE, payload: item };
+};
